changed datatype of total_spent

+++ config/db.py
@@ -22,7 +22,7 @@ def init_db():
         login_id INT NOT NULL,\
         name VARCHAR(30),\
         budget INT DEFAULT 0,\
-        total_spent INT DEFAULT 0,\
+        total_spent NUMERIC DEFAULT 0,\
         phone VARCHAR(30),\
         profession VARCHAR(30),\
         alert BOOLEAN DEFAULT FALSE,\
